MAR-246 - sync the healthcheck app with the api, this endpoint is used by pingdom
