+++ healthcheck/constants.py
@@ -1,0 +1,3 @@
+class HealthStatus:
+    OK = "OK"
+    FAIL = "FAIL"
