+++ healthcheck/middleware.py
@@ -1,12 +1,15 @@
 import time
 
 
-class HealthCheckMiddleware(object):
-    def __init__(self, get_response):
-        self.get_response = get_response
-
-    def __call__(self, request):
+class StatsMiddleware:
+    def process_request(self, request):
+        """ Start time at request coming in """
         request.start_time = time.time()
-        response = self.get_response(request)
 
+    def process_response(self, request, response):
+        """ End of request, take time """
+        total = time.time() - request.start_time
+
+        # Add the header.
+        response["X-Response-Time-Duration-ms"] = int(total * 1000)
         return response
