+++ healthcheck/urls.py
@@ -1,8 +1,15 @@
 from django.urls import path
-from healthcheck.views import HealthCheckView
+from django.utils.decorators import decorator_from_middleware
+
+from .middleware import StatsMiddleware
+from .views import HealthCheckView
 
 app_name = "healthcheck"
 
 urlpatterns = [
-    path("check/", HealthCheckView.as_view(), name="check"),
+    path(
+        "check/",
+        decorator_from_middleware(StatsMiddleware)(HealthCheckView.as_view()),
+        name="check"
+    ),
 ]
