+++ tests/healthcheck/test_middleware.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 from django.test import TestCase
 
-from healthcheck.middleware import HealthCheckMiddleware
+from healthcheck.middleware import StatsMiddleware
 
 
 class TestMiddleware(TestCase):
@@ -13,11 +13,11 @@ class TestMiddleware(TestCase):
 
     def setUp(self):
         get_response = Mock()
-        self.middleware = HealthCheckMiddleware(get_response)
+        self.middleware = StatsMiddleware()
         self.request = Mock(spec=[""])
 
     def test_middleware_start_time_added(self):
         """ Checks start_time is added to a request object in middleware"""
         self.assertFalse(hasattr(self.request, "start_time"))
-        self.middleware(self.request)
+        self.middleware.process_request(self.request)
         self.assertTrue(hasattr(self.request, "start_time"))
